# openapi_debug.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Primus Partners Portal API",
    "description": "Core API for Dynamics-linked portal",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify Login",
        "operationId": "verify_login_auth_login_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login/resend-otp": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Resend Otp",
        "operationId": "resend_otp_auth_login_resend_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendOtpDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forgot Password",
        "operationId": "forgot_password_auth_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify Otp",
        "operationId": "verify_otp_auth_verify_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset Password",
        "operationId": "reset_password_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/meetings/free-slots": {
      "post": {
        "tags": [
          "Meetings"
        ],
        "summary": "Get free slots of project managers",
        "description": "Pass one or more project-manager emails and a date range. Returns computed free time windows (within business hours 9 AM \u2013 6 PM) for each person so the caller can pick an available slot.",
        "operationId": "get_free_slots_meetings_free_slots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FreeSlotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FreeSlotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/meetings/schedule": {
      "post": {
        "tags": [
          "Meetings"
        ],
        "summary": "Schedule a new Teams meeting",
        "description": "Creates a Teams online meeting for the specified organizer. Set `enforce_availability` to true to block scheduling when any attendee is busy during the requested time slot.",
        "operationId": "schedule_meeting_meetings_schedule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/meetings/my-meetings": {
      "get": {
        "tags": [
          "Meetings"
        ],
        "summary": "List meetings created by the current user",
        "description": "Returns meeting records stored in the database for the authenticated user.",
        "operationId": "list_my_meetings_meetings_my_meetings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DateTimeWithZone": {
        "properties": {
          "dateTime": {
            "type": "string",
            "title": "Datetime"
          },
          "timeZone": {
            "type": "string",
            "title": "Timezone",
            "default": "India Standard Time"
          }
        },
        "type": "object",
        "required": [
          "dateTime"
        ],
        "title": "DateTimeWithZone",
        "description": "ISO-8601 datetime with timezone identifier."
      },
...
      "ForgotPasswordDTO": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "type": {
            "type": "string",
            "enum": [
              "admin",
              "advisor",
              "alumni",
              "vendor",
              "client"
            ],
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "email",
          "type"
        ],
        "title": "ForgotPasswordDTO"
      },
      "FreeSlotRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "schedules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Schedules"
          },
          "startTime": {
            "$ref": "#/components/schemas/DateTimeWithZone"
          },
          "endTime": {
            "$ref": "#/components/schemas/DateTimeWithZone"
          },
          "availabilityViewInterval": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 5.0,
            "title": "Availabilityviewinterval",
            "default": 30
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "schedules",
          "startTime",
          "endTime"
        ],
        "title": "FreeSlotRequest",
        "description": "Request body for fetching free slots of project managers.\n- user_id: the caller's Graph user ID or UPN (used as context for getSchedule)\n- schedules: list of project-manager emails/UPNs whose calendars to check\n- startTime / endTime: the date-range window to search\n- availabilityViewInterval: granularity in minutes (default 30)"
      },
      "FreeSlotResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/PersonFreeSlots"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "FreeSlotResponse",
        "description": "Aggregated response containing free slots per person."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginDTO": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "type": {
            "type": "string",
            "enum": [
              "admin",
              "advisor",
              "alumni",
              "vendor",
              "client"
            ],
            "title": "Type"
          },
          "captcha_token": {
            "type": "string",
            "title": "Captcha Token"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "type",
          "captcha_token"
        ],
        "title": "LoginDTO"
      },
      "LoginResponseDTO": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "vendor_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Type"
          },
          "vendor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Name"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "LoginResponseDTO"
      },
      "MeetingCreateRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "startDateTime": {
            "type": "string",
            "title": "Startdatetime"
          },
          "endDateTime": {
            "type": "string",
            "title": "Enddatetime"
          },
          "attendees": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attendees"
          },
          "enforce_availability": {
            "type": "boolean",
            "title": "Enforce Availability",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "subject",
          "startDateTime",
          "endDateTime"
        ],
        "title": "MeetingCreateRequest",
        "description": "Request body to schedule a new Teams meeting.\n- user_id: organizer's Graph user ID or UPN\n- subject: meeting title\n- startDateTime / endDateTime: ISO-8601 strings\n- attendees: list of attendee email addresses (optional)\n- enforce_availability: if True, the backend checks all attendees are free\n  before creating the meeting; returns 409 if anyone is busy"
      },
...
      "PersonFreeSlots": {
        "properties": {
          "scheduleId": {
            "type": "string",
            "title": "Scheduleid"
          },
          "freeSlots": {
            "items": {
              "$ref": "#/components/schemas/TimeSlot"
            },
            "type": "array",
            "title": "Freeslots"
          }
        },
        "type": "object",
        "required": [
          "scheduleId",
          "freeSlots"
        ],
        "title": "PersonFreeSlots",
        "description": "Free slots for one person."
      },
...
      "ResendOtpDTO": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "type": {
            "type": "string",
            "enum": [
              "admin",
              "advisor",
              "alumni",
              "vendor",
              "client"
            ],
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "email",
          "type"
        ],
        "title": "ResendOtpDTO"
      },
      "ResetPasswordDTO": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "reset_token": {
            "type": "string",
            "title": "Reset Token"
          },
          "new_password": {
            "type": "string",
            "maxLength": 12,
            "minLength": 8,
            "title": "New Password"
          },
          "confirm_password": {
            "type": "string",
            "maxLength": 12,
            "minLength": 8,
            "title": "Confirm Password"
          },
          "type": {
            "type": "string",
            "enum": [
              "admin",
              "advisor",
              "alumni",
              "vendor",
              "client"
            ],
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "email",
          "reset_token",
          "new_password",
          "confirm_password",
          "type"
        ],
        "title": "ResetPasswordDTO"
      },
...
      "TimeSlot": {
        "properties": {
          "start": {
            "type": "string",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end"
        ],
        "title": "TimeSlot",
        "description": "A single free window."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyOtpDTO": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "otp": {
            "type": "string",
            "maxLength": 6,
            "minLength": 4,
            "title": "Otp"
          },
          "type": {
            "type": "string",
            "enum": [
              "admin",
              "advisor",
              "alumni",
              "vendor",
              "client"
            ],
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "email",
          "otp",
          "type"
        ],
        "title": "VerifyOtpDTO"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}